refactor(backend): split app/server and load env natively without dotenv
Move listen + graceful shutdown into a new server.js; app.js stays as the
exported Express app so tests import it without opening a port.

Drop the dotenv dependency: dev/start/test scripts load the root .env via
node --env-file-if-exists, and prisma.config.js uses process.loadEnvFile.
Remove a dead #config/loadEnv.js import in seed.js that broke pnpm
seed/setup, and route the root seed through prisma db seed (which resolves
.env by absolute path, cwd-independent).
