+++ backend/Dockerfile
@@ -65,4 +65,4 @@ EXPOSE 8000
 HEALTHCHECK --interval=30s --timeout=5s --start-period=30s --retries=3 \
   CMD wget -qO- http://localhost:8000/health >/dev/null 2>&1 || exit 1
 ENTRYPOINT ["/sbin/tini", "--"]
-CMD ["node", "src/app.js"]
+CMD ["node", "src/server.js"]
