+++ backend/jest.teardown.js
@@ -6,12 +6,10 @@
  * No corre si jest es interrumpido con Ctrl-C; para ese caso queda como
  * fallback restart del backend container.
  *
- * NOTA: globalTeardown corre como módulo separado y NO aplica el
- * setupFiles: ['dotenv/config'] del jest.config. Cargamos dotenv aquí
- * para que prisma.js encuentre DATABASE_URL.
+ * Las env vars (DATABASE_URL) las carga el flag --env-file-if-exists del
+ * script de test en el proceso principal de jest, que globalTeardown hereda.
  */
 
-import 'dotenv/config'
 import { prisma } from '#config/prisma.js'
 
 export default async function teardown() {
