+++ backend/package.json
@@ -3,11 +3,11 @@
   "version": "1.0.0",
   "main": "server.js",
   "scripts": {
-    "test": "node --experimental-vm-modules ./node_modules/jest/bin/jest.js --forceExit",
-    "test:debug": "node --experimental-vm-modules ./node_modules/jest/bin/jest.js --forceExit --detectOpenHandles",
-    "file": "node --experimental-vm-modules ./node_modules/jest/bin/jest.js --forceExit --testPathPatterns",
-    "start": "node src/app.js",
-    "dev": "prisma generate && node --watch-path=src --watch-path=../shared src/app.js",
+    "test": "node --experimental-vm-modules --env-file-if-exists=../.env ./node_modules/jest/bin/jest.js --forceExit",
+    "test:debug": "node --experimental-vm-modules --env-file-if-exists=../.env ./node_modules/jest/bin/jest.js --forceExit --detectOpenHandles",
+    "file": "node --experimental-vm-modules --env-file-if-exists=../.env ./node_modules/jest/bin/jest.js --forceExit --testPathPatterns",
+    "start": "node --env-file-if-exists=../.env src/server.js",
+    "dev": "prisma generate && node --env-file-if-exists=../.env --watch-path=src --watch-path=../shared src/server.js",
     "prisma:pull": "prisma db pull",
     "prisma:generate": "prisma generate",
     "prisma:sync": "pnpm run prisma:pull && pnpm run prisma:generate",
@@ -20,7 +20,6 @@
     "@prisma/client": "7.4.2",
     "bcryptjs": "3.0.3",
     "cors": "2.8.5",
-    "dotenv": "17.2.3",
     "express": "5.2.1",
     "express-rate-limit": "8.4.1",
     "express-session": "1.19.0",
