+++ backend/src/config/prisma.js
@@ -1,4 +1,3 @@
-import './loadEnv.js'
 import { PrismaClient } from '@prisma/client'
 import { PrismaMariaDb } from '@prisma/adapter-mariadb'
